+++ src/lib/components/Forms/InputWrapper.svelte
@@ -11,18 +11,23 @@
 	export let iconText: string | null = null;
 	export let pending: boolean = false;
 	export let pendingMessage: string[] = [];
-	export let errors: string[] | [];
+	export let errors: string[] | [] = [];
 	export let isValid: boolean;
 	export let tooltip: string | undefined = undefined;
 	export let statusIcons = true;
 	export let disabled = false;
+	export let required = false;
+	export let labelColor = 'var(--clr-text-main)';
 </script>
 
 {#if label}
-	<label for={name} class="row-2" class:disabled>
+	<label for={name} class="row-2" class:disabled style={`color: ${labelColor}`}>
 		{label}
 		{#if tooltip}
 			<TooltipIcon {tooltip} width={0.75} />
+		{/if}
+		{#if required}
+			<span class="required">*</span>
 		{/if}
 	</label>
 {/if}
@@ -63,7 +68,7 @@
 
 <style type="scss">
 	label {
-		padding-bottom: 0.2em;
+		padding-bottom: 6px;
 		align-items: flex-start;
 	}