+++ src/routes/+page.svelte
@@ -1,4 +1,5 @@
 <script type="ts">
+	import DropZone from './../lib/components/DropZone/DropZone.svelte';
 	import InputWrapper from './../lib/components/Forms/InputWrapper.svelte';
 	import Range from './../lib/components/Range/Range.svelte';
 	import Label from '$lib/components/Label/Label.svelte';
@@ -7,6 +8,7 @@
 	import PoweredByEcdao from '$lib/components/PoweredByECDAO/PoweredByECDAO.svelte';
 	import StatusCircle from '$lib/components/StatusCircle/StatusCircle.svelte';
 	import { Tab, TabPanel, TabList, Tabs } from '$lib/index';
+	import FlowConnection from '$lib/components/FlowConnection/FlowConnection.svelte';
 	import Accordion from '$lib/components/Accordion/Accordion.svelte';
 
 	let stepss = [
@@ -25,16 +27,29 @@
 	];
 
 	let value = 5;
+	let file: File[];
 </script>
 
 <section>
-	<div class="container-small column-6 align-center">
-		<label for="a">
+	<div class="container-small column-6">
+		<a href="dwd" class="card-primary">
+			<h1>Card</h1>
+		</a>
+		<!-- <label for="a">
 			<input type="checkbox" id="a" />
 			Hola
-		</label>
+		</label> -->
 
-		<Label color="alert" size="xx-small">Holacomova</Label>
+		<FlowConnection network="testnet" transactionInProgress={true} />
+
+		<label for="logo">Logo</label>
+		<DropZone name="logo" accept="image/png" maxAmountOfFiles={1} bind:bindValue={file} />
+
+		<InputWrapper name="hola" isValid={true} label="algo">
+			<input type="text" name="hola" required />
+		</InputWrapper>
+
+		<!-- <Label color="alert" size="xx-small">Holacomova</Label>
 
 		<label for="burn-tokens" class="switch">
 			<input type="checkbox" name="burn-tokens" id="burn-tokens" placeholder="e.g. 1.000.000" />
@@ -150,6 +165,7 @@
 				<h4>Third panel</h4>
 			</TabPanel>
 		</Tabs>
+	</div> -->
 	</div>
 </section>
 <section class="container">
@@ -171,6 +187,7 @@
 </section>
 <PoweredByEcdao />
 
+<!-- <PoweredByEcdao /> -->
 <style type="scss">
 	h1 {
 		--font-weight: var(--font-weight-semibold);